Caredata-Visualization-Azure/Backend/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
passlib[bcrypt]>=1.7.4
azure-data-tables>=12.4.0
openai>=1.0.0
